fix(task): tasks re-attach to ProviderSyncbackRequests
Summary:
The old task queue for message changing tasks (like mark unread and
archiving) had a system that would only do 5 at a time until the Promises
resolved. Unfortunately our new system doesn't resolve the request
promises until the full IMAP task is actually complete. When you're moving
dozens of messages this can take a very long time since they're currently
not prioritized in sync.

Furthermore if you quit the app while this is inprogress, the Tasks had no
way of saving the status of each of the syncback actions. This means when
you restarted the app, the TaskQueue would get re-inflated and try tasks
again. We now keep track of any tasks that have started and either look
for them in the database or re-listen for their changing status again.

Test Plan: manual

Reviewers: mark, halla, juan

Reviewed By: juan

Differential Revision: https://phab.nylas.com/D3666

+++ src/flux/tasks/destroy-category-task.es6
@@ -53,17 +53,26 @@ export default class DestroyCategoryTask extends Task {
     // delta which comes after a delay
     NylasAPI.incrementRemoteChangeLock(Category, this.category.serverId);
 
-    return new SyncbackTaskAPIRequest({
-      api: NylasAPI,
-      options: {
-        accountId,
-        path,
-        method: 'DELETE',
-        returnsModel: false,
-      },
-    })
-    .run()
-    .thenReturn(Task.Status.Success)
+    let runPromise = Promise.resolve();
+
+    if (this._syncbackRequestId) {
+      runPromise = SyncbackTaskAPIRequest.waitForQueuedRequest(this._syncbackRequestId)
+    } else {
+      runPromise = new SyncbackTaskAPIRequest({
+        api: NylasAPI,
+        options: {
+          accountId,
+          path,
+          method: 'DELETE',
+          returnsModel: false,
+          onSyncbackRequestCreated: (syncbackRequest) => {
+            this._syncbackRequestId = syncbackRequest.id
+          },
+        },
+      }).run()
+    }
+
+    return runPromise.thenReturn(Task.Status.Success)
     .catch(APIError, (err) => {
       if (!NylasAPI.PermanentErrorCodes.includes(err.statusCode)) {
         return Promise.resolve(Task.Status.Retry);
